I changed the app name in package.json
I created a ProductList folder. Inside the folder, I created a three
different components. 1st component is ProductList.js: This component
is responsible for rendering product list. 2nd component is Available
Products which allow me to create DUMMY_PRODUCTS data in order to
display them inside my product list component. This application does not
fetch any data from the database so it is simply set in the Available
product list component.

The last component is for  summary about the products which is rendered
inside the product list component.

Finally I rendered ProductList component inside the app.js

I render AvailableProducts and ProductListSummary inside the ProductList
component.

I created css file for AvailableProducts but it is not completed yet.

I created css file for Product Summary and it is completed.

I added the "app" next to Grocery Shopping in h1 tag. that is why header
component is changed.